requirements.txt
colorlog==5.0.1
firebase_admin==6.1.0
Flask==2.2.3
Flask_RESTful==0.3.9
langchain==0.0.129
openai==0.27.0
promptlayer==0.1.79
requests==2.28.1
SocksiPy_branch==1.01
tiktoken==0.3.0
